requirements.txt
anthropic>=0.25.0
streamlit>=1.35.0
python-pptx>=0.6.23
openpyxl>=3.1.2
python-docx>=1.1.2
pdfplumber>=0.11.0
chromadb>=0.5.0
python-dotenv>=1.0.1
pandas>=2.2.0
Pillow>=10.3.0
tiktoken>=0.7.0
